try to fix colab

+++ main.py
@@ -70,7 +70,7 @@ def trainEncoder(generator, train_ds, epochs=40, save_img_every=100):
             if train_counter % save_img_every == 0:
                 with train_summary_writer.as_default():
                     image_idx = 0
-                    image_input = tf.Variable([images[image_idx]])
+                    image_input = tf.Variable([images[image_idx]], dtype=tf.float64)
                     tf.summary.image(
                         "generator_img",
                         generator(image_input),
@@ -106,7 +106,6 @@ def Q1(epochs=10, save_img_every=100):
 
 def visual_latent_space(generator: exModels.CNNGenerator, test_ds):
     pass
-
 
     # from sklearn.manifold import TSNE
     # from sklearn.decomposition import PCA,LatentDirichletAllocation
